Adjustments to book search to add back button while keeping search results
